Reduce the size of a Bit from 16 to 2 bytes

+++ cli/liblisa-semantics-tool/src/merge.rs
@@ -64,7 +64,7 @@ impl Merge {
                                         let mut indices_to_remove = Vec::new();
                                         let bit_indices = encoding.bits.iter()
                                             .enumerate()
-                                            .filter(|&(_, &bit)| bit == Bit::Part(part_index))
+                                            .filter(|&(_, &bit)| bit == Bit::Part(part_index as u8))
                                             .map(|(index, _)| index)
                                             .collect::<Vec<_>>();
                                         let mut n = 0;
